Fix prefetching from http/https sources
Files from the web (https/http) had no thumbs in the gallery, as the workdirectory was prepended to the url. By checking whether the image link is a weblink, this can be restricted to local files only (if this exists, maybe also look for ftp:// etc.).
I am not fluent in LUA (this is basically my first code in that language), so feel free to refactor to your liking.