Use real youtube thumbnails from youtube-dl
gallery-thumbgen.lua checks to see if the input_path starts with
https?://.  If so, it uses youtube-dl to attempt to resolve a url to a
thumbnail (i.e. the thumbnail used on the website.)  This works for any
site for which youtube-dl can find thumbnails.

In the special case of a link to youtube proper, youtube-dl is not used
because it's slow.  Instead the URL is parsed to pull out the video ID
and a URL to the thumbnail for that particular ID is generated.  This is
much faster than waiting for youtube-dl, but perhaps more brittle.

If any of the above fails, then the original input_path is returned and
thumbnailing proceeds as usual.