preserve (and optionally display) playlist titles
playlists have an optional 'title' subproperty that is used, for
example, by ytdl_hook.lua.  This title subproperty of playlists is only
populated if present in the playlist file, so to preserve it I've
modified restore_playlist_and_select to create an M3U playlist in
memory.

I've also added a simple option to display these playlist titles instead
of filenames, which creates a nice UX when viewing youtube playlists.